refactor: split package of config
